Refactored tests to allow console run
