+++ test/guinness_test.dart
@@ -1,48 +1,16 @@
 library guinness_test;
 
 import 'package:guinness/guinness_html.dart' as guinness;
-import 'package:unittest/unittest.dart';
-import 'package:dartmocks/dartmocks.dart';
-import 'dart:html' as html;
-import 'dart:async';
-import 'html_utils_test.dart' as htmlUtilsTest;
 
-part 'src/syntax_test.dart';
-part 'src/model_test.dart';
-part 'src/spy_test.dart';
-part 'src/integration_test.dart';
-part 'src/exclusive_visitor_test.dart';
-part 'src/unittest_backend_test.dart';
-part 'src/suite_info_test.dart';
+import 'common_test.dart' as common;
+import 'html/html_utils_test.dart' as htmlUtilsTest;
+import 'html/unittest_backend_test.dart' as unittest_backend;
 
-class DummyContext implements guinness.Context {
-  withDescribe(guinness.Describe describe, Function definition) {}
-}
-
-final context = new DummyContext();
-noop() {}
-
-createSuite() => new guinness.Suite(context);
-
-createDescribe({bool exclusive: false, bool excluded: false, parent,
-    Function func: noop, String name: ""}) => new guinness.Describe(
-    name, parent, context, func, exclusive: exclusive, excluded: excluded);
-
-createIt({bool exclusive: false, bool excluded: false, parent,
-    Function func: noop, String name: ""}) => new guinness.It(
-    name, parent, func, exclusive: exclusive, excluded: excluded);
-
-main() {
+void main() {
   guinness.guinnessEnableHtmlMatchers();
   guinness.guinness.autoInit = false;
 
-  testIntegration();
-  testExclusiveVisitor();
-  testSuiteInfo();
-  testSyntax();
-  testModel();
-  testSpy();
-  testUnitTestBackend();
-
+  common.main();
   htmlUtilsTest.main();
+  unittest_backend.main();
 }
